requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.100.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0